requirements.txt
fastapi==0.95.1
uvicorn==0.21.1
pika==1.2.1
motor==3.1.1
sqlalchemy==2.0.20
asyncpg==0.27.0
cx_Oracle==8.3.0
psycopg2-binary==2.9.5
python-dotenv